More tests, moving fixtures to dedicated conftest module
